### 20250524, v0.9.1 Build 530 * CLANG devel     * ANSI C apps no printf() output with **clang-cl** tool chain

+++ toroCLibrary/LibCore/_cdeMemStrxCpy.c
@@ -49,7 +49,7 @@ void* _cdeMemStrxCpy(IN int preset, OUT void* pszDest, IN const void* pszSource,
     //
     // strncpy support -- that fills remaining buffer until count
     //
-    if ((-1 != count) && PRESET_BOZ) {
+    if (((size_t) -1 != count) && PRESET_BOZ) {
         for ( /*n = 0*/; n < count; n++, pDst16 = &pDst16[xcrement], pDst8 = &pDst8[xcrement]) {
 
             if (PRESET_WID) {
