+++ toroCLibrary/Library/stdio_h/Ungetc.c
@@ -61,7 +61,7 @@ Returns
 
 int ungetc(int c, FILE* stream) {
     CDEFILE* pCdeFile = (CDEFILE*)stream;
-    unsigned int nRet = (size_t)EOF;
+    unsigned nRet = (unsigned)EOF;
 
     do {
